migration tweak

+++ backend/src/db/migrations/20250429232917_store-cert-secret-key-and-chain.ts
@@ -3,7 +3,10 @@ import { Knex } from "knex";
 import { TableName } from "../schemas";
 
 export async function up(knex: Knex): Promise<void> {
-  if (await knex.schema.hasTable(TableName.CertificateBody)) {
+  if (
+    (await knex.schema.hasTable(TableName.CertificateBody)) &&
+    !(await knex.schema.hasColumn(TableName.CertificateBody, "encryptedCertificateChain"))
+  ) {
     await knex.schema.alterTable(TableName.CertificateBody, (t) => {
       t.binary("encryptedCertificateChain").nullable();
     });
@@ -25,7 +28,10 @@ export async function down(knex: Knex): Promise<void> {
     await knex.schema.dropTable(TableName.CertificateSecret);
   }
 
-  if (await knex.schema.hasTable(TableName.CertificateBody)) {
+  if (
+    (await knex.schema.hasTable(TableName.CertificateBody)) &&
+    (await knex.schema.hasColumn(TableName.CertificateBody, "encryptedCertificateChain"))
+  ) {
     await knex.schema.alterTable(TableName.CertificateBody, (t) => {
       t.dropColumn("encryptedCertificateChain");
     });
